Add a fade transition, transition target, and basic appearance functionality.
This PR includes the first modular transition instance, `FadeTransition`. This sets the foundation for a catalog of transition types.

Includes unit tests but no examples yet.

By default, `FadeTransition` will fade in or out the fore view of the transition depending on the direction of the transition. This makes it easy to create a one-liner "fade" transition:

```swift
viewController.transitionController.transition = FadeTransition()
```

`FadeTransition` is also meant to be used as a modular part of a larger transition.

For example, say we wanted to fade a background element out:

```swift
final class MyCustomTransition: Transition {
  func start(with context: TransitionContext) {
    // ...other logic...

    let fadeTransition = FadeTransition(target: .target(view: backgroundView))
    fadeTransition.appearance = .disappearing
    context.compose(with: fadeTransition)

    // ...other logic...
  }
}
```

`FadeTransition` will mirror the transition values when the view controller is dismissed.

`MDMTransitionTarget` is essentially an enum type that can be one of three values:

1. The fore view.
2. The back view.
3. A custom view.

This object allows a transition builder to refer to the back and fore fixture views without having access to the view instances (commonly the case when assigning a Transition type to the transition controller).

`MDMTransitionAppearance` defines whether a view is appearing or disappearing.

`MDMTransitionTimingMode` is an enum of three possible timing mode interpretations:

1. Mirrored. The default.
2. Relative.
3. Absolute.

When mirrored, timing information is treated as relative and the animation values are automatically reversed when the transition is animating backward. This means a view that fades in during the last half of a presentation transition will fade **out** during the **first** half of the dismissal transition.

When relative, timing information is treated as relative and animation values are not reversed.

When absolute, timing information is treated as absolute and animation values are not reversed.